backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
gunicorn
psycopg2-binary
"celery[redis]"
django-celery-beat
pillow
requests
imagehash

django-environ
